ワーニング削除

+++ src/core/environ/win32/Application.cpp
@@ -557,7 +557,7 @@ bool tTVPApplication::ProcessMessage( MSG &msg ) {
 				DispatchMessage(&msg);
 			}
 		} else {
-			TVPTerminateCode = msg.wParam;
+			TVPTerminateCode = (int)msg.wParam;
 			tarminate_ = true;
 		}
 	}
